face_recognition_project/requirements.txt
# Face Recognition Project Dependencies
# Core ML and Computer Vision
torch>=1.11.0
torchvision>=0.12.0
opencv-python>=4.5.0
Pillow>=8.3.0

# Face Recognition Libraries
insightface>=0.6.0
facenet-pytorch>=2.5.0
dlib>=19.22.0

# Data Science and Visualization
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Utilities
tqdm>=4.62.0
pyyaml>=6.0
colorama>=0.4.4

# Report Generation
reportlab>=3.6.0
jinja2>=3.0.0

# Development
pytest>=6.0.0
black>=21.0.0